Move getFQDNOrHostName to base/common/
