+++ base/common/CMakeLists.txt
@@ -6,6 +6,7 @@ set (SRCS
     DateLUT.cpp
     DateLUTImpl.cpp
     demangle.cpp
+    getFQDNOrHostName.cpp
     getMemoryAmount.cpp
     getThreadId.cpp
     JSON.cpp
@@ -20,15 +21,9 @@ set (SRCS
 )
 
 if (ENABLE_REPLXX)
-    set (SRCS ${SRCS}
-        ReplxxLineReader.cpp
-        ReplxxLineReader.h
-    )
+    list (APPEND SRCS ReplxxLineReader.cpp)
 elseif (ENABLE_READLINE)
-    set (SRCS ${SRCS}
-        ReadlineLineReader.cpp
-        ReadlineLineReader.h
-    )
+    list (APPEND SRCS ReadlineLineReader.cpp)
 endif ()
 
 if (USE_DEBUG_HELPERS)
@@ -56,8 +51,6 @@ if(CCTZ_LIBRARY)
     target_link_libraries(common PRIVATE ${CCTZ_LIBRARY})
 endif()
 
-target_link_libraries(common PUBLIC replxx)
-
 # allow explicitly fallback to readline
 if (NOT ENABLE_REPLXX AND ENABLE_READLINE)
     message (STATUS "Attempt to fallback to readline explicitly")
@@ -82,11 +75,13 @@ endif ()
 
 target_link_libraries (common
     PUBLIC
+        ${Poco_Net_LIBRARY}
         ${Poco_Util_LIBRARY}
         ${Poco_Foundation_LIBRARY}
         ${CITYHASH_LIBRARIES}
         ${Boost_SYSTEM_LIBRARY}
         FastMemcpy
+        replxx
 )
 
 if (ENABLE_TESTS)
