+++ dbms/src/Client/ConnectionPoolWithFailover.cpp
@@ -4,7 +4,7 @@
 #include <Poco/Net/DNS.h>
 
 #include <Common/BitHelpers.h>
-#include <Common/getFQDNOrHostName.h>
+#include <common/getFQDNOrHostName.h>
 #include <Common/isLocalAddress.h>
 #include <Common/ProfileEvents.h>
 #include <Core/Settings.h>
