+++ dbms/src/Functions/FunctionFQDN.cpp
@@ -1,7 +1,7 @@
 #include <Functions/IFunctionImpl.h>
 #include <Functions/FunctionFactory.h>
 #include <DataTypes/DataTypeString.h>
-#include <Common/getFQDNOrHostName.h>
+#include <common/getFQDNOrHostName.h>
 #include <Core/Field.h>
