+++ dbms/src/Interpreters/ClientInfo.cpp
@@ -4,7 +4,7 @@
 #include <IO/ReadHelpers.h>
 #include <IO/WriteHelpers.h>
 #include <Core/Defines.h>
-#include <Common/getFQDNOrHostName.h>
+#include <common/getFQDNOrHostName.h>
 #include <Common/ClickHouseRevision.h>
 #include <Common/config_version.h>
 #include <unistd.h>
